Fix entryGuardRegisterConnectStatus and document existing heuristics
entryGuardRegisterConnectStatus would not mark guards we failed to connect to
and never made contact before as unreachable.

The existing heuristic for tor in this case is removing the guard from the
guard list to make sure it will not be tried again, but in our case simply
marking as unreachable should be enough.

It also documents the existing heuristics in
entry_guard_register_connect_status().